e3-causes-of-overhead/cprofile/requirements.txt
numpy==2.1.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
jinja2==3.1.4